Rename: web - api, controller 디렉토리 구분
api: restful,
controller: mvc
